requirement.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.5
commonmark==0.9.1
Deprecated==1.2.13
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-rest-auth==0.9.5
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
humanize==4.0.0
idna==2.10
keyboard==0.13.5
Pillow==9.1.1
pycparser==2.21
PyGithub==1.55
Pygments==2.12.0
PyJWT==2.4.0
PyNaCl==1.5.0
python-decouple==3.6
python-gitlab==3.3.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.1
requests==2.28.0
requests-toolbelt==0.9.1
reviews==0.4.2
rich==12.2.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.1
